requirements.txt
fastapi
uvicorn
openai
sqlalchemy
psycopg2-binary
google-generativeai